refactor(auth): Phase 4 move token services to Domain/Tokens/
- Relocate JwtTokenService, ServiceTokenService and their
  ITokenService/IServiceTokenService interfaces to Domain/Tokens/
  (namespace Auth.Service.Domain.Tokens); public signatures unchanged.
- Update consumers (LoginHandler, ServiceTokenEndpoint,
  TokenStartupExtensions) and affected tests to the new namespace;
  ServiceTokenService keeps a using for slice-local ServiceClientOptions
  (moves with the IssueServiceToken slice in phase 6).
