+++ auth-microservice/Auth.Service/Domain/Tokens/ServiceTokenService.cs
@@ -3,11 +3,12 @@ using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Text;
 using Auth.Service.Domain;
+using Auth.Service.Services;
 using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
 using Microsoft.IdentityModel.Tokens;
 
-namespace Auth.Service.Services;
+namespace Auth.Service.Domain.Tokens;
 
 public class ServiceTokenService : IServiceTokenService
 {
