+++ auth-microservice/Auth.Tests/LoginHandlerTests.cs
@@ -1,6 +1,7 @@
 using System.Security.Cryptography;
 using Auth.Service.Domain;
 using Auth.Service.Domain.Abstractions;
+using Auth.Service.Domain.Tokens;
 using Auth.Service.Services;
 using Auth.Service.Services.Signing;
 using ECommerce.Shared.Authentication;
